Develop

+++ api/internal/users/application/useCases/deleteUsecase.go
@@ -1,6 +1,7 @@
 package usecases
 
 import (
+	"errors"
 	sv "suffgo/internal/shared/domain/valueObjects"
 	"suffgo/internal/users/domain"
 )
@@ -15,7 +16,11 @@ func NewDeleteUsecase(repository domain.UserRepository) *DeleteUsecase {
 	}
 }
 
-func (s *DeleteUsecase) Execute(id sv.ID) error {
+func (s *DeleteUsecase) Execute(id sv.ID, CurrentUserID sv.ID) error {
+
+	if id != CurrentUserID {
+		return errors.New("unauthorized")
+	}
 
 	err := s.userDeleteRepository.Delete(id)
